requirements.txt
langchain-community
langchain-openai
faiss-cpu
openai
beautifulsoup4
requests
pandas
openpyxl
python-docx
markdown
markdownify
python-frontmatter
mistune
autoscraper
scrapy
selenium
jsonschema
ebooklib
tifffile
opencv-python
rasterio
Pillow
pillow-heif
tensorflow
numpy
ultralytics
torch
torchvision
cmake
deepface
deepface
tf-keras
transformers
diffusers
gtts
extract_msg
openai-whisper
psycopg2
pypdf
pdfplumber
playwright
pyocr
pypdf[crypto]
llama-index
pymupdf4llm

